Expose last issuance timestamp via expvar.
This commit adds a `expvar.Int` published under "lastIssuance" that
contains the timestamp of the last successful certificate issuance. This
allows easy creation of a script that monitors the debug server to
ensure that there has been a successful issuance within a set period
(e.g. last five minutes).

Solves #1945.