Do not invoke SA for identical reg updates.
When the RA's `UpdateRegistration` function is called by the WFE it
would always pass the update to the SA via `ra.SA.UpdateRegistration`,
even when there was no difference between the existing registration and
the proposed update.

This commit modifies `core/objects.go`'s implementation of the
`Registration` `MergeUpdate` function such that it can indicate via
a boolean return value whether the merge actually changed the base
object or not.

The RA is also updated to examine the boolean return from `MergeUpdate`.
If the merge didn't produce a change then the RA will return involving
the SA, saving some needless database work.