Fix index out of range panic in `getCerts`.
Boulder issue #2004 describes a panic observed in `getCerts` caused by
an index out of range. It appears as though this is caused by a race
condition between the initial `SelectOne` lookup for the count of
certificates, and the subsequent individual `Select` queries to fetch
the Certificates. If the number of eligible certificates changes between
these points (e.g. due to certificates expiring) there is a potential
that one of the Select calls will return an empty result set. If this
happens, then the `lastSerial` update will access an index out of range.

This commit adds an explicit `len` check to the processing loop before
the `lastSerial` update. If there are no results returned from the DB
query then the loop is broken.

A test case for the fix was written and included in this PR. The
testcase initially caused the out of range panic observed in #2004.
After adding the `len` fix in this commit the test began passing without
error.