`-end` bugfix, example config files, documentation.
This commit fixes a bug with the `-end` parameter where the default
(99999999) would cause a slice index out of range error. This was fixed
by setting the -end value to `len(m.destinations)` in `run` when it is
too large.

This commit also adds a separated config examplex for the
notify-mailer into `test/config` and `test/config-next` respectively.

Finally a documentation string was added to describe the overall design
& usage of the notification mailer.