Separate Contact equality check, safe nil handling.
Based on review feedback this commit moves the `AcmeURL` equality
checking out of `MergeUpdate` into a dedicated `contactsEqual` function.
This makes `MergeUpdate` much easier to read. In the future once #1966
is merged we can replace much of `contactsEqual` with simple string
comparisons.

This commit also fixes a nil dereference bug when an update is posted
with a JSON "null" value in the Contact array. The `contactsEqual`
method now explicitly checks each Contact slice index to make sure it is
not null before calling `String`. We allow a `null` value to be
considered an update that `MergeUpdate` will allow because the
validation step that throws out the nils and produces a user error is
surfaced later post-merge by the RA calling `validateContacts`.