Change `DBConfig`'s `URL` func to strip whitespace.
As described in issue #2025 if the contents of a `DBConnectFile`
argument ends in a trailing newline an error is produced when later
using the contents as a DB configuration URL.

This PR changes the `URL` function of the `DBConfig` type to strip
leading and trailing whitespace (including newlines) when reading the
`DBConnectFile` contents. This resolves #2025.