Fix dbConnect strings in OCSP Responder.
Right now we use the Source field for both DB and file URLs. However, we want to
move to the DBConnect config field, so that we can take advantage of the code
that reads DSNs from a file on disk. It turns out the existing code didn't work
if you configure a dbConnect string, because it would error out with:

"source" parameter not found in JSON config

After rearranging, both methods should work.