Remove empty reg update optimization.
For #2001 an optimization was added to the WFE to avoid invoking the
RA's `UpdateRegistration` method when a trivial (e.g.
`{"resource:"reg"}`) update is received. Instead the WFE returned the
trivial update back to the client immediately.

This is contrary to the ACME spec which indicates:
  Servers SHOULD NOT respond to GET requests for registration resources
  as these requests are not authenticated. If a client wishes to query
  the server for information about its account (e.g., to examine the
  “contact” or “certificates” fields), then it SHOULD do so by sending
  a POST request with an empty update. That is, it should send a JWS
  whose payload is trivial ({“resource”:”reg”}).

The optimization regression was captured in issue #2066 when it broke at
least one client implementation.

This removes the empty reg update optimization and passes all POST's to
the RA. The RA will in turn fetch the existing registration to return to
the client. The second half of the #2001 optimizations remains in place,
no DB UPDATE's will be performed if the new registration content
doesn't differ from the existing registration content (as determine by
the return of `registration.MergeUpdate`.

Since the WFE optimization is no longer in place the
`FailureRegistrationAuthority` mock isn't required and is removed.
Similarly `TestEmptyRegistration` in `wfe_test.go` is changed from
testing for the optimization to testing for the ACME described "get
registration info" behaviour.