Remove `--config` for mail-test-srv.
The `start()` function of `startservers.py` was providing a `--config`
argument to the `mail-test-srv` binary. Prior to +997d439 this command
did *not* call `flag.Parse()`, even though it defined a `-http` flag at
the top of the file. In effect this meant that all of the CLI flags
given to the command were ignored and never parsed. Post +997d439 this
caused `startservers.py` to encounter a "flag provided but not defined:
--config" error when starting the `mail-test-srv`.

Since the config was never honoured I have elected to remove the
`--config` from `startservers.py` rather than add a `--config` flag to
the server.