Add `--closeChance` parameter to test flaky server.
This commit adds a `--closeChance` flag that allows specifying a chance
between 0 and 100%. After every received `MAIL` command the server
obtains a random uint between 0 and 100. If this uint is < the provided
close chance then the connection is terminated and no MAIL processing
occurs.

This parameter allows testing reconnect logic in integration tests by
having some number of client `MAIL` messages result in the server
terminating the connection.