Fix errors in unit tests, mutex pass by value.
This commit fixes the unit tests that were broken by the mail server refactor.
Mostly this meant instantiating a copy of a `mailSrv` and using this
instead of global var/func references.

This commit also fixes the receiver for the `mailSrv` functions to use
a pointer to a `mailSrv` instance. Without doing this the `mailSrv` is
being passed by value, which is a big no-no for structs with a mutex in
them (like the `mailSrv`).