Create rpm during CI run.
This commit modifies the `test.sh` script to allow a `rpm` value in the
`RUN` parameter passed to the script via the environment. When present,
`make rpm` is invoked and a good status is required for the build to
pass.

The `Makefile` was modified to add a `-f` to the `fpm` invocation used
by the `rpm` build task to allow the output rpm to be overwritten if
present. Otherwise multiple runs of identical bulld (e.g. on a local dev
machine) would collide on the .rpm already being present.

Finally `.travis.yml` is updated to include `rpm` in the `RUN` used
during CI such that an RPM is built by default for CI runs.

This fixes #2085