Abort reconnecting after a max num of attempts.
Previous to this commit if the remote mailserver never recovered we
would exponentially backoff more and more but would always continue
trying to reconnect.

This commit introduces a new parameter to control the maximum # of
attempts. If this maximum is reached the EOF from the remote mailserver
is returned to the client and we do not attempt a reconnect.

The two mailer commands are edited to have the "-retryAttempts" flag to
wire this maximum to the mailer instaces.