Review feedback pt1.
This commit renames the `retryBase` and `retryMax` parameters in both
the mailer and the two consuming commands. The new names are
`reconnectBase` and `reconnectMax` to be more explicit about their
purpose. The default `reconnectMax` value for the `expiration-mailer`
and `notify-mailer` was lowered to 5m.

This commit also removes `retryAttempts` from the mailer and commands.
Its intended purpose is duplicated with `reconnectMax` and we use that
one elsewhere so dropping `retryAttempts` makes the most sense. With
`retryAttempts` removed there was no purpose to the
`TestReconnectFailure()` unit test and it was removed as well.