Add `NotAfter` field to `CertificateStatus`.
As part of #1864 we identified that a `JOIN` in two of the most
expensive queries from the slow query log could be avoided if the
`certificateStatus` table had a `NotAfter` field with the `ExpiresAt`
data from the `certificate` table.

This commit introduces a DB migration to add this field as well as the
required modification to the `core.CertificateStatus` object to hold it.
The SA is updated to populate this field when new certificates are
created.

To handle back-filling the existing rows of the `certificateStatus` table a new `notafter-backfiller` utility is provided. After being used to backfill the existing rows it can be deleted having served its purpose (off to the great util farm in the sky for retirement).