Remove `SELECT *` usage from OCSP updater.
Both the `findStaleOCSPResponses` and `findRevokedCertificatesToUpdate` functions used a `SELECT *` on the `certificateStatus` table. When the `notAfter` field is `NULL` (e.g. immediately after the migration and before the `notafter-backfiller` cmd has run) the `ocsp-updater` cmd produces:

```
  ocsp-updater [AUDIT] Failed to find stale OCSP responses: sql:
  Scan error on column index 9: unsupported Scan, storing driver.Value
  type <nil> into type *time.Time
```

Best practice is to avoid `SELECT *` and so this commit replaces the `*`
in each function with the fields that end up being accessed by the
callers.

The alternative fix would be to use an empty `time.Time` struct value as
the `certificateStatus` `notAfter` field instead of `NULL` but this
feels icky overall.