Adds `$GOPATH` troubleshooting to docker quickstart.
We have seen a couple issues from folks that run into trouble using
`docker-compose up` with invalid `$GOPATH`'s configured (e.g. see
issues  #2150, #2141, #2112).

This commit adds a sentence to the README indicate that if you see
a docker "oci runtime error" or a failure to create the container it may
be caused by your `$GOPATH` and to check that first.