Adds new `notAfter` and `isExpired` fields to certStatus.
This commit adds a migration to create two new fields on the
`certificateStatus` table: `notAfter` and `isExpired`. The rationale for
these fields is explained in #1864[0].

The change is primarily in three components:

a) the `CertificateStatus` struct in `core/objects.go` is updated to
have members for the two new fields. Without these Gorp will error once
the table structure is migrated.

b) the `NewCertificate` function in `sa/sa.go` is updated to save
a value in the `NotAfter` member going forward for new certificates
created after this point.

c) `cmd/ocsp-updater/main.go` is updated in two places to replace
`SELECT *` usage for the `certificateStatus` table.

Note: Updating the `isExpired` field in the OCSP updater can not be
done yet, the `notAfter` field needs to be fully populated first.
Similarly we can't remove the `JOIN` on `certificates` from the
`findStaleOCSPResponse` query until all DB rows have `notAfter`
populated.

[0] - https://github.com/letsencrypt/boulder/issues/1864