Moves `AddTableWithName` calls, makes bool flag global.
Cleaning up past mistakes and moving the `AddTableWithName` and
`SetKeys` calls on the dbmap to the `initTables` func in
`sa/database.go` with the others.

Unfortunately Gorp throws:
```
panic: No ColumnMap in table certificateStatus type
       certStatusModelv2 with field Serial
```
if there are table mappings made for both `certStatusModelv1` and
  `certStatusModelv2` at the same time. So, we need to conditionally add
one map or the other based on the config flag...

Except the config flag was a field of the SQLStorageAuthority and not
visible to the stand-alone `initTables` func from `database.go`. The
cleanest solution to this appears to me to be making the flag bool
a global, so that's what I did :'(