Fixes `certStatusModelv2` struct embedding.
Previously the inner `certStatusModelv1` defined in the
`certStatusModelv2` struct was a pointer. This isn't necessary and was
causing Gorp to err on insert being unable to convert the pointer type
to something it knew how to map. This commit makes the embedded type
literal (is that the right term? hmm) and not a pointer.