Retires `LookupIPv6` VA flag.
The `LookupIPv6` flag has been enabled in production and isn't required
anymore. This commit removes the flag entirely.

The `errA` and `errAAAA` error handling in `LookupHost` is left *as-is*
- meaning that a non-nil `errAAAA` will *not* be returned to the caller.
This matches the existing behaviour, and the expectations of the
`TestDNSLookupHost` unit tests.

This commit also removes the tests from `TestDNSLookupHost` that tested
the `LookupIPv6 == false` behaviours since those are no longer
implemented.