Reworks existing DB check, always runs `goose`.
Prior to this commit an optimization was present in `test/create_db.sh`
that would `exit 0` if the `boulder_sa_integration` database existed.
This early exit meant that after the first invocation of `create_db.sh`
no further `goose` migrations would be applied unless the operator
dropped their databases or edited the script.

This commit reworks the existing DB optimization so that it only skips
the `CREATE DATABASE` statements and allows `goose` to try and apply
migrations.

This doesn't result in significantly longer start up times because Goose
is smart enough to know when no migrations are required and outputs
something similar to:
  `goose: no migrations to run. current version: 20160602142227`

This should address #2174.