Fixes empty `./sa/_db-next/migrations` problem.
Goose exits non-zero if there are no migrations to apply with the error
message:
   "no valid version found"

This commit addresses that for the initial state where we have no "next"
migrations yet by using `find` to check that the path exists, is
a directory, and is not empty. If these conditions are not met we don't
invoke `goose` at all.