Adds test to ensure consistency w/ `pendingStatuses`.
This commit follows on the previous to add a protective unit test to
ensure that `CountPendingAuthorizations` and the `pendingStatus` list
remain in-sync.

The `TestPendingStatusList` unit tests checks the `pendingStatuses` list
against the statuses used by `CountPendingAuthorization` to ensure they
are equal. If they aren't, an error is raised instructing the programmer
that updated `pendingStatuses` that they also need to update
`CountPendingAuthorization` and `TestPendingStatusList`.