Handles legacy pendingAuthorization delete carefully.
Prior to this commit when `FinalizeAuthorization` was dealing with
a legacy pending authz (e.g. a pending authz in the
`pendingAuthorizations` table vs a pending authz in the `authz` table)
it used `tx.Delete` to remove the old row after creating the new `authz`
row. It did this without checking the # of affected rows.

In this commit `FinalizeAuthorization` is changed to use `tx.Exec` and
to explicitly check the `sql.Result`'s number of affected rows to ensure
it is exactly 1.

Using `tx.Delete` means we don't need a throw-away
`pendingauthzModel{}`. Checking the # of affected rows means this is
overall a safer operation and we ensure expectations match reality.